Fixed cryptic undefined bug
When attempting to load EXIF data on an image before DOM ready, two
messages were being sent to the global JS file. From there, it seems as
if one of them was being prevented by the cross-domain XHR restriction
and the other was working fine. Not totally sure what the issue was,
but by only allowing EXIF lookups after DOM ready, the problem is
prevented.